+++ figpack-figure/src/figpack-main-plugin/shared/component-time-scroll-view-3/TimeScrollToolbar.tsx
@@ -74,6 +74,8 @@ const TimeScrollToolbar: FunctionComponent<Props> = ({
     [onInteractionModeChange],
   );
 
+  console.log("--- on zoom y");
+
   return (
     <div
       style={{